Fixed an ios9 problem.  Also fixed a few compiler warnings that only turn up with certain settings.
Former-commit-id: 084274f4eceb252fa296f7383c1206db73dcc798

+++ ios/library/WhirlyGlobeLib/include/GeometryManager.h
@@ -113,7 +113,7 @@ public:
     /// The triangles, which reference points
     std::vector<RawTriangle> triangles;
     /// A texture ID for the geometry
-    int texId;
+    long long texId;
 };
 
 /// Represents a single Geometry Instance
